Syntax 0.1.10: Shared Attributes + Date
Implement new `attributes` context shared by all nodes, regardless of
which attributes are supported by each node. This approach is not only
easier to maintain, but also makes the syntax more error tolerant, which
means that it will handle better copy-and-paste operations, when users
move around snippets to work quicker. It's not the job of the highlighter
to validate the document, hence it should highlight attributes and their
value regardless of the node.

But the new system will only match attributes immediately following an
opening tag or a valid attribute definition, or on a new line after `\`;
so the system discerns the appropriate context for attributes.

Adapt `[doc`, `[ch` and `unknown_node` contexts to use the new shared
attributes context.

Implement `date` attribute, with a temporary scoping name for the date
format/value, until I can come up with better scoping semantics.