Add Inline Code
Implement the `[c` node. Unlike its font siblings `[b` and `[i`, this
element also scopes its contents to allow custom styles. In order to
do so properly, any leading whitespace is consumed and not scoped as
part of the contents, even if these are split across multiple lines,
indented. This mimics the real life behaviour of pmlc.