Key-Shortcuts: Bold & Italic
Implement shortcuts ALT+B and ALT+I to wrap selection in `[b` and `[i`
respectively. Based on the MarkdownEditing package implementation, which
provides some smart checks to handle correctly single and multi cursors
with and without selections, according to context.

https://github.com/SublimeText-Markdown/MarkdownEditing

Bold and italic formatting are disabled if a cursor/selection is inside
and `[html` node.