Amend Installation Instructions
Clarify in "Installation" section that the package won't self-update.

Add new "Package Roadmap" explaining how in the nearby future the focus
is going to shift toward developing a package for VSCode, which will
take precedence over this package; and how the ultimate goal is the
creation of a PML Lang Server, which will replace this package.