Drop Versioning & Releases
After realizing that with the current installation method
(i.e. from GH repo instead of Package Control channel)
the package gets updated whenever there are new commits
in `main`, instead of only when there's a tagged release,
and that as a result of this update messages were not being
shown to end users, we decided to stop versioning the package
and creating tagged releases (see #31).

* Delete previous package update messages, keeping only the
  installation message. Update `messages/README.md` accordingly.
* Edit `CHANGELOG.md` explaining why the package won't be any
  longer versioned and subject to tagged releases.
  Substitute all entries titles by using dates instead of
  release version numbers.
* Add new menu "Tools > PML > Sublime PML Changelog" to open
  the repository CHANGELOG in the browser.
* User Guide:
    * Add to preamble a TIP reminding users to check the
      CHANGELOG, pointing to the new menu to launch it.
    * Remove package versioning, just keep the date.